// src/schemas/auth-schema.ts
import { z } from "zod";

export const authSchema = z.object({
  email: z.string().email({ message: "Invalid email address" }),
  password: z
    .string()
    .min(6, { message: "Password must be at least 6 characters" }),
  first_name: z.string().optional(),
  last_name: z.string().optional(),
});

export type AuthFormData = z.infer<typeof authSchema>;